fix: track algod spec instead of pinning a stale one

+++ cmd/bootstrap.go
@@ -22,7 +22,7 @@ import (
 const CheckAlgodInterval = 10 * time.Second
 const CheckAlgodTimeout = 2 * time.Minute
 
-var CatchpointLagThreshold int = 30_000
+var CatchpointLagThreshold uint64 = 30_000
 
 // bootstrapCmdShort provides a brief description of the "bootstrap" command to initialize a fresh Algorand node.
 var bootstrapCmdShort = "Initialize a fresh node"
